refactor(config): migrate to antd
