+++ src/pages/Config/components/WhatsAppInvitationLinkInput/index.css
@@ -1,0 +1,11 @@
+.valid-whatsapp-icon {
+  color: green;
+}
+
+.invalid-whatsapp-icon {
+  color: red;
+}
+
+.whatapp-invitation-url-container input {
+  width: 400px;
+}
